chore: 20230612

+++ src/array/minimum_number_of_arrows_to_burst_balloons.rs
@@ -13,6 +13,6 @@ impl Solution {
                 endpoint = point[1]
             }
             count
-        }) as i32
+        })
     }
 }
